refactor(organisms): merge SfContentPages
