--- packages/vue/src/components/organisms/SfContentPages/SfContentPages.html
@@ -1,52 +1,0 @@
-<div class="sf-content-pages">
-  <SfBar
-    :title="active ? active : title"
-    :back="isActive"
-    class="mobile-only"
-    @click:back="updatePage()"
-  />
-  <section
-    class="sf-content-pages__section"
-    :class="{'sf-content-pages__section--is-active': isActive}"
-  >
-    <div class="sf-content-pages__sidebar">
-      <h1 class="sf-content-pages__title desktop-only">{{title}}</h1>
-      <div v-for="(category, key) in categories" :key="key">
-        <h2 v-if="category.title" class="sf-content-pages__category-title">
-          {{category.title}}
-        </h2>
-        <SfList class="sf-content-pages__list">
-          <SfListItem
-            v-for="page in category.items"
-            :key="page.title"
-            class="sf-content-pages__list-item"
-          >
-            <!-- @slot Custom menu-item markup -->
-            <slot name="menu-item" v-bind="{updatePage, page, active}">
-              <SfMenuItem
-                @click="updatePage(page.title)"
-                :label="page.title"
-                class="sf-content-pages__menu"
-                :class="{'sf-menu-item--is-active': page.title === active}"
-              >
-                <template #icon>
-                  <SfIcon
-                    v-if="page.icon"
-                    size="xxs"
-                    color="#686d7d"
-                    :icon="page.icon"
-                    class="sf-menu-item__icon"
-                  />
-                </template>
-              </SfMenuItem>
-            </slot>
-          </SfListItem>
-        </SfList>
-      </div>
-    </div>
-    <div class="sf-content-pages__content">
-      <!-- @slot Slot for Content Page -->
-      <slot />
-    </div>
-  </section>
-</div>
